moved item related typings to types/item
