+++ src/types/index.ts
@@ -1,20 +1,4 @@
-export interface ItemNotification {
-  id: string;
-  date: Date;
-}
-
-export interface ItemCreation {
-  description: string;
-  amount: number;
-  months: number[];
-  notification?: ItemNotification;
-}
-
-export interface Item extends ItemCreation {
-  id: string;
-  isPaid: boolean;
-  notification: ItemNotification;
-}
+import { Item } from "./item";
 
 export interface StorageData {
   amountLeft: number;
